Move definition of cmd args to cli.arguments
