--- rasa/core/cli/__init__.py
@@ -1,5 +1,0 @@
-import rasa.core.cli.arguments
-import rasa.core.cli.test
-import rasa.core.cli.run
-import rasa.core.cli.train
-import rasa.core.cli.visualization
