+++ rasa/core/train.py
@@ -1,5 +1,3 @@
-import argparse
-import asyncio
 import logging
 import os
 import typing
@@ -11,44 +9,6 @@ if typing.TYPE_CHECKING:
     from rasa.core.interpreter import NaturalLanguageInterpreter
 
 logger = logging.getLogger(__name__)
-
-
-def create_argument_parser():
-    """Parse all the command line arguments for the training script."""
-    from rasa.core import cli
-
-    parser = argparse.ArgumentParser(
-        description="Train a dialogue model for Rasa Core. "
-        "The training will use your conversations "
-        "in the story training data format and "
-        "your domain definition to train a dialogue "
-        "model to predict a bots actions."
-    )
-    parent_parser = argparse.ArgumentParser(add_help=False)
-    cli.train.add_general_args(parent_parser)
-
-    subparsers = parser.add_subparsers(help="Training mode of core.", dest="mode")
-    subparsers.required = True
-
-    train_parser = subparsers.add_parser(
-        "default", help="train a dialogue model", parents=[parent_parser]
-    )
-    compare_parser = subparsers.add_parser(
-        "compare",
-        help="train multiple dialogue models to compare policies",
-        parents=[parent_parser],
-    )
-    interactive_parser = subparsers.add_parser(
-        "interactive",
-        help="teach the bot with interactive learning",
-        parents=[parent_parser],
-    )
-
-    cli.train.add_compare_args(compare_parser)
-    cli.train.add_interactive_args(interactive_parser)
-    cli.train.add_train_args(train_parser)
-
-    return parser
 
 
 async def train(
@@ -100,15 +60,6 @@ async def train(
     agent.persist(output_path, dump_stories)
 
     return agent
-
-
-def _additional_arguments(args):
-    additional = {
-        "augmentation_factor": args.augmentation,
-        "debug_plots": args.debug_plots,
-    }
-    # remove None values
-    return {k: v for k, v in additional.items() if v is not None}
 
 
 async def train_comparison_models(
@@ -172,19 +123,6 @@ async def get_no_of_stories(story_file, domain):
         story_file, TemplateDomain.load(domain)
     )
     return len(stories)
-
-
-async def do_default_training(cmdline_args, stories, additional_arguments):
-    """Train a model."""
-
-    await train(
-        domain_file=cmdline_args.domain,
-        stories_file=stories,
-        output_path=cmdline_args.out,
-        dump_stories=cmdline_args.dump_stories,
-        policy_config=cmdline_args.config[0],
-        kwargs=additional_arguments,
-    )
 
 
 async def do_compare_training(cmdline_args, stories, additional_arguments):
