+++ setup.py
@@ -65,7 +65,7 @@ install_requires = [
     "gevent~=1.4",
     "pytz~=2018.9",
     "python-dateutil~=2.7",
-    "rasa_core_sdk~=1.0.0a1",
+    "rasa_core_sdk~=0.13.1",
     "colorclass~=2.2",
     "terminaltables~=3.1",
     "sanic~=18.12.0",
